fix: breaking changes from upstream
Breaking changes in v30 leads to breaking changes in this helm chart. By
default renovate does not look anymore after config.json files. Only if
RENOVATE_CONFIG_FILE environment is defined and points to a config file
in json format.

To fix the changes and alternatively support config files in javaScript
format, appConfig became a dictonary to specify the config as well as
the file extension.